feat: move power up to hive power session
